Start to design API
A simple start of the design of the API.
The current thinking is to use an interface to represent the different types
of sensor to provide abstraction from the sensor drivers.

These interfaces should return readings (also an interface) that provide data
formatting. E.g. The `temperatureSensor` interface would return a temperature
interface that has methods like `inKelvin()`, `inCelsius()`.

This would mean that a valid API call would be similar to:
`temperatureSensor.getTemperature().inCelsius();`